Service fixes and code clean up

+++ src/main/groovy/com/electro/bikeapp/domains/BikeDomain.groovy
@@ -38,10 +38,10 @@ class BikeDomain {
     String color
 
     @Column(name = 'wholesale_cost')
-    BigDecimal wholesaleCost
+    Float wholesaleCost
 
     @Column(name = 'retail_cost')
-    BigDecimal retailCost
+    Float retailCost
 
     @Column(name = 'description')
     String description
