+++ src/main/groovy/com/electro/bikeapp/dtos/AddProductDTO.groovy
@@ -12,9 +12,9 @@ class AddProductDTO {
 
     String color
 
-    BigDecimal wholesaleCost
+    Float wholesaleCost
 
-    BigDecimal retailCost
+    Float retailCost
 
     String description
