+++ src/main/groovy/com/electro/bikeapp/dtos/VendorOrderDTO.groovy
@@ -20,6 +20,6 @@ class VendorOrderDTO {
 
     Integer quantityReceived
 
-    BigDecimal purchasePrice
+    Float purchasePrice
 
 }
